Added C# UI, added GIF demo
* All files relevant to the C# app have been added to the repo

* A GIF has been created to demonstrate the UI

* The README doc has been amended to reflect these additions
