+++ src/csharp/Home.xaml
@@ -1,0 +1,67 @@
+﻿<Page x:Class="F1_Strategy_Interface.Home"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:F1_Strategy_Interface"
+      mc:Ignorable="d" 
+      d:DesignHeight="600" d:DesignWidth="1020"
+      Title="F1 Strategy Interface - Home">
+
+    <Page.Background>
+        <ImageBrush ImageSource="intro_background.jpg" Opacity="1"/>
+    </Page.Background>
+
+    <Page.Resources>
+        <Style TargetType = "Button">
+            <Setter Property="Background" Value="SlateBlue" />
+            <Setter Property="BorderThickness" Value="2.5" />
+            <Setter Property="BorderBrush" Value="Black" />
+            <Setter Property="FontFamily" Value="Verdana" />
+            <Setter Property="FontSize" Value="20" />
+            <Setter Property="Foreground" Value="White" />
+            <Setter Property="FontWeight" Value="Bold" />
+            <Setter Property="Width" Value="417" />
+        </Style>
+        <Style TargetType = "Label">
+            <Setter Property="Foreground" Value="Black" />
+            <Setter Property="FontFamily" Value="Verdana" />
+            <Setter Property="FontSize" Value="20" />
+            <Setter Property="VerticalAlignment" Value="Center" />
+            <Setter Property="HorizontalAlignment" Value="Center" />
+            <Setter Property="FontWeight" Value="Bold" />
+        </Style>
+    </Page.Resources>
+
+    <Grid Margin="50,30,50,30">
+
+        <Grid.Background>
+            <SolidColorBrush Color="White" Opacity="0.85"/>
+        </Grid.Background>
+        
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="*" />
+            <ColumnDefinition Width="*" />
+        </Grid.ColumnDefinitions>
+
+        <Grid.RowDefinitions>
+            <RowDefinition Height="20"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="10"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="10"/>
+            <RowDefinition Height="Auto"/>
+        </Grid.RowDefinitions>
+
+        <Label Grid.Column="0" Grid.Row="1" Grid.ColumnSpan="2" FontSize="32">
+            The Strategy Hub
+        </Label>
+
+        <Button Grid.Column="0" Grid.Row="3" x:Name="Button_Plot_Prev_Race" Click ="Button_Plot_Prev_Race_Click" Content="Plot historical race lap times" />
+        <Button Grid.Column="1" Grid.Row="3" x:Name="Button_Find_Optimal_Pit_Laps" Click="Button_Plot_Optimal_Pit_Laps_Click" Content="Plot optimal race strategies" />
+
+        <Image Grid.Column="0" Grid.Row="5" Source="plot_laptimes.png" Height="300"/>
+        <Image Grid.Column="1" Grid.Row="5" Source="plot_optimal_strategies.png" Height="300"/>
+
+    </Grid>
+</Page>
