+++ src/csharp/Intro.xaml
@@ -1,0 +1,72 @@
+﻿<Page x:Class="F1_Strategy_Interface.Intro"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:F1_Strategy_Interface"
+      mc:Ignorable="d" 
+      d:DesignHeight="600" d:DesignWidth="1020"
+      Title="F1 Strategy Hub - Welcome">
+    
+    <Page.Resources>
+        <Style TargetType = "Button">
+            <Setter Property="Background" Value="SlateBlue" />
+            <Setter Property="BorderThickness" Value="2.5" />
+            <Setter Property="BorderBrush" Value="Black" />
+            <Setter Property="FontFamily" Value="Verdana" />
+            <Setter Property="FontSize" Value="20" />
+            <Setter Property="Foreground" Value="White" />
+            <Setter Property="FontWeight" Value="Bold" />
+            <Setter Property="Width" Value="140" />
+        </Style>
+        <Style TargetType = "Label">
+            <Setter Property="Foreground" Value="Black" />
+            <Setter Property="FontFamily" Value="Verdana" />
+            <Setter Property="FontSize" Value="20" />
+            <Setter Property="VerticalAlignment" Value="Center" />
+            <Setter Property="HorizontalAlignment" Value="Center" />
+            <Setter Property="FontWeight" Value="Bold" />
+        </Style>
+    </Page.Resources>
+
+    <Page.Background>
+        <ImageBrush ImageSource="intro_background.jpg" Opacity="1"/>
+    </Page.Background>
+
+    <Grid Margin="10,10,10,10">
+        
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="*" />
+        </Grid.ColumnDefinitions>
+
+        <Grid.RowDefinitions>
+            <RowDefinition Height="20"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="20"/>
+            <RowDefinition Height="40"/>
+            <RowDefinition Height="*"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="20"/>
+        </Grid.RowDefinitions>
+
+        <Label Grid.Column="0" Grid.Row="1" FontSize="48">
+            Welcome to the
+        </Label>
+
+        <Label Grid.Column="0" Grid.Row="2" FontSize="48">
+            Strategy Hub
+        </Label>
+
+        <Button Grid.Column="0" Grid.Row="4" x:Name="Button_Intro" Click="Button_Click" Content="Continue" />
+
+        <Label Grid.Column="0" Grid.Row="6" FontSize="22" Background="White">
+            Developed by Jamie Hilton
+        </Label>
+
+        <Label Grid.Column="0" Grid.Row="7" FontSize="22" Background="White">
+            v1.2 (November 2021)
+        </Label>
+    </Grid>
+</Page>
