+++ src/csharp/MainWindow.xaml
@@ -1,0 +1,9 @@
+﻿<NavigationWindow x:Class="F1_Strategy_Interface.MainWindow"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:F1_Strategy_Interface"
+        mc:Ignorable="d"
+        Title="F1 Strategy Interface" Height="600" Width="1020" ShowInTaskbar="False" Source="Intro.xaml">
+</NavigationWindow>
