+++ src/csharp/Task_PlotPreviousRaceLapTimes.xaml
@@ -1,0 +1,119 @@
+﻿<Page x:Class="F1_Strategy_Interface.Task_PlotPreviousRaceLapTimes"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:F1_Strategy_Interface"
+      mc:Ignorable="d" 
+      d:DesignHeight="600" d:DesignWidth="1020"
+      Title="Task_PlotPreviousRaceLapTimes">
+
+    <Page.Resources>
+        <Style TargetType = "Button">
+            <Setter Property="Background" Value="SlateBlue" />
+            <Setter Property="BorderThickness" Value="2.5" />
+            <Setter Property="BorderBrush" Value="Black" />
+            <Setter Property="FontFamily" Value="Verdana" />
+            <Setter Property="FontSize" Value="20" />
+            <Setter Property="Foreground" Value="White" />
+            <Setter Property="FontWeight" Value="Bold" />
+            <Setter Property="Width" Value="200" />
+        </Style>
+        <Style TargetType = "Label">
+            <Setter Property="Foreground" Value="Black" />
+            <Setter Property="FontFamily" Value="Verdana" />
+            <Setter Property="FontSize" Value="20" />
+            <Setter Property="VerticalAlignment" Value="Center" />
+            <Setter Property="HorizontalAlignment" Value="Center" />
+            <Setter Property="FontWeight" Value="Bold" />
+        </Style>
+    </Page.Resources>
+
+    <Page.Background>
+        <ImageBrush ImageSource="intro_background.jpg" Opacity="1"/>
+    </Page.Background>
+
+    <Grid Margin="50,30,50,150" x:Name="Container">
+
+        <Grid.Background>
+            <SolidColorBrush Color="White" Opacity="0.85"/>
+        </Grid.Background>
+
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="*" />
+            <ColumnDefinition Width="*" />
+        </Grid.ColumnDefinitions>
+
+        <Grid.RowDefinitions>
+            <RowDefinition Height="20"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="20"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="20"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="20"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="20"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="30"/>
+            <RowDefinition Height="Auto"/>
+            <RowDefinition Height="20"/>
+        </Grid.RowDefinitions>
+
+        <Label Grid.Column="0" Grid.Row="1" Grid.ColumnSpan="2" FontSize="32">
+            Plot historical race lap times
+        </Label>
+
+        <Border Grid.Column="0" Grid.Row="4" Height="26">
+            <Label VerticalAlignment="Top" HorizontalAlignment="Right" Foreground="Black" FontSize="14">Select a year:</Label>
+        </Border>
+        
+        <ComboBox Grid.Column="1" Grid.Row="4" Name="yrListBox" Width="150" HorizontalAlignment="Left">
+            <ComboBoxItem>2019</ComboBoxItem>
+        </ComboBox>
+
+        <Border Grid.Column="0" Grid.Row="6" Height="26">
+            <Label VerticalAlignment="Top" HorizontalAlignment="Right" Foreground="Black" FontSize="14">Select a race:</Label>
+        </Border>
+
+        <ComboBox Grid.Column="1" Grid.Row="6" Name="rndListBox" Width="150" HorizontalAlignment="Left">
+            <ComboBoxItem>2: Bahrain Grand Prix</ComboBoxItem>
+            <ComboBoxItem>9: Austrian Grand Prix</ComboBoxItem>
+            <ComboBoxItem>12: Hungarian Grand Prix</ComboBoxItem>
+            <ComboBoxItem>14: Italian Grand Prix</ComboBoxItem>
+        </ComboBox>
+
+        <Border Grid.Column="0" Grid.Row="8" Height="26">
+            <Label VerticalAlignment="Top" HorizontalAlignment="Right" Foreground="Black" FontSize="14">Select a driver:</Label>
+        </Border>
+
+        <ComboBox Grid.Column="1" Grid.Row="8" Name="driverListBox" Width="150" HorizontalAlignment="Left">
+            <ComboBoxItem>Alexander Albon</ComboBoxItem>
+            <ComboBoxItem>Antonio Giovinazzi</ComboBoxItem>
+            <ComboBoxItem>Carlos Sainz</ComboBoxItem>
+            <ComboBoxItem>Charles Leclerc</ComboBoxItem>
+            <ComboBoxItem>Daniel Ricciardo</ComboBoxItem>
+            <ComboBoxItem>Daniil Kvyat</ComboBoxItem>
+            <ComboBoxItem>George Russell</ComboBoxItem>
+            <ComboBoxItem>Kevin Magnussen</ComboBoxItem>
+            <ComboBoxItem>Kimi Räikkönen</ComboBoxItem>
+            <ComboBoxItem>Lance Stroll</ComboBoxItem>
+            <ComboBoxItem>Lando Norris</ComboBoxItem>
+            <ComboBoxItem>Lewis Hamilton</ComboBoxItem>
+            <ComboBoxItem>Max Verstappen</ComboBoxItem>
+            <ComboBoxItem>Nico Hülkenberg</ComboBoxItem>
+            <ComboBoxItem>Pierre Gasly</ComboBoxItem>
+            <ComboBoxItem>Robert Kubica</ComboBoxItem>
+            <ComboBoxItem>Romain Grosjean</ComboBoxItem>
+            <ComboBoxItem>Sebastian Vettel</ComboBoxItem>
+            <ComboBoxItem>Sergio Pérez</ComboBoxItem>
+            <ComboBoxItem>Valtteri Bottas</ComboBoxItem>
+        </ComboBox>
+
+        <Button Grid.Column="0" Grid.Row="10" Grid.ColumnSpan="2"  x:Name="Button_Run" Click ="ButtonRunClick" Content="Run" />
+        <TextBlock Grid.Column="0" Grid.Row="11" Grid.ColumnSpan="2"  x:Name="Message_Box" HorizontalAlignment="Center" FontSize="14" TextWrapping="Wrap" Text=""/>
+        <Button Grid.Column="0" Grid.Row="12" Grid.ColumnSpan="2"  x:Name="Button_RetHome" Click="ButtonRetHomeClick" Content="Home" />
+
+    </Grid>
+</Page>
