+++ src/csharp/bat_files/run_python_plot_historical_race_lap_times.bat
@@ -1,0 +1,2 @@
+cd "C:\Users\jamie\OneDrive\Documents\F1 Strategy Model"
+"C:\Users\jamie\OneDrive\Documents\F1 Strategy Model\venv\Scripts\python.exe" "C:\Users\jamie\OneDrive\Documents\F1 Strategy Model\c_sharp_app_plot_historical_race_lap_times.py" "%~1" "%~2" "%~3"
