+++ src/python/c_sharp_app_plot_historical_race_lap_times.py
@@ -1,0 +1,9 @@
+import modelling_utilities as mu
+import sys
+
+yr = int(sys.argv[1])
+rnd = int(sys.argv[2])
+dvr = sys.argv[3]
+
+results_df = mu.download_laptimes(yr, rnd, dvr)
+mu.plot_laptimes(results_df)
